refactoring
move implements to site-checker.ts
